[SECURITY] Prevent authentication of user from other service
Ensure the "authUser" step in the authentication process
only verdicts positive on user data discovered by this very
service and not by any other auth service.

Relying on the 'tx_oidc' field of the user data is not
sufficient, as any user previously logged in has a value in
this field.

We secure this step by preserving the Access Token, supplied
by the actual OpenID connect authentication process during
TYPO3's "getUser" step, in the user's data. If this token is
detected the in the "authUser" step, we got an actual login
by our service.