[TASK] Move AuthenticationController logic to middleware
The code in the AuthenticationController does not
need anything from Extbase.
In fact, it was only triggered by a custom PAGE TypoScript
object, which itself was only triggered
by the callback middleware.

To simplify this code, the whole code is moved directly
to the middleware, removing all the Extbase overhead.