[!!!][TASK] Remove implicit authenticationContext from OpenIdConnectService
Ensure that the OpenIdConnectService does not store the authentication
context. Create and pass the authentication context explicitly where
needed.

Set the authentication context cookie only when redirecting to the
(external) authorization endpoint of the identity provider.

[!!!] Mark the login_url parameter for the (internal) authentication
endpoint as required. Without a login_url the callback middleware does
not know where to send the user to login.

Remove the redirect url from the authorization context. It can still be
passed to the (internal) authentication endpoint and will be
incorporated into the login_url.